Use Tanstack data for tx amounts and fee worth
