+++ apps/mobile-wallet/src/features/send/screens/TotalWorthRow.tsx
@@ -1,35 +1,18 @@
-import { AddressHash, AssetAmount, calculateTokenAmountWorth, selectAllPrices } from '@alephium/shared'
-import { ALPH } from '@alephium/token-list'
-import { useMemo } from 'react'
+import { AssetAmount } from '@alephium/shared'
+import { useFetchTokensAmountsWorth } from '@alephium/shared-react'
 import { useTranslation } from 'react-i18next'
 
 import Amount from '~/components/Amount'
 import Row from '~/components/Row'
-import { useAppSelector } from '~/hooks/redux'
-import { makeSelectAddressesKnownFungibleTokens } from '~/store/addresses/addressesSelectors'
-import { getTransactionAssetAmounts } from '~/utils/transactions'
 
 interface TotalWorthRowProps {
   assetAmounts: AssetAmount[]
-  fromAddress: AddressHash
 }
 
-// This component will be improved by reusing the worth in desktop wallet after migrating to Tanstack
-const TotalWorthRow = ({ assetAmounts, fromAddress }: TotalWorthRowProps) => {
-  const tokenPrices = useAppSelector(selectAllPrices)
-  const selectAddressesKnownFungibleTokens = useMemo(() => makeSelectAddressesKnownFungibleTokens(), [])
-  const knownFungibleTokens = useAppSelector((s) => selectAddressesKnownFungibleTokens(s, fromAddress))
+const TotalWorthRow = ({ assetAmounts }: TotalWorthRowProps) => {
   const { t } = useTranslation()
 
-  const { attoAlphAmount, tokens } = getTransactionAssetAmounts(assetAmounts)
-  const assets = [{ id: ALPH.id, amount: attoAlphAmount }, ...tokens]
-
-  const totalWorth = assets.reduce((totalWorth, token) => {
-    const tokenInfo = knownFungibleTokens.find(({ id }) => id === token.id)
-    const tokenPrice = tokenPrices.find(({ symbol }) => symbol === tokenInfo?.symbol)?.price
-
-    return totalWorth + calculateTokenAmountWorth(BigInt(token.amount), tokenPrice ?? 0, tokenInfo?.decimals ?? 0)
-  }, 0)
+  const { data: totalWorth } = useFetchTokensAmountsWorth(assetAmounts)
 
   if (!totalWorth) return null
