+++ packages/shared-react/src/api/apiDataHooks/transaction/index.ts
@@ -2,3 +2,5 @@ export * from '@/api/apiDataHooks/transaction/useFetchPendingTransaction'
 export * from '@/api/apiDataHooks/transaction/useFetchTransaction'
 export * from '@/api/apiDataHooks/transaction/transactionTypes'
 export * from '@/api/apiDataHooks/transaction/useFetchTransactionTokens'
+export * from '@/api/apiDataHooks/transaction/useFetchTokensAmountsWorth'
+export * from '@/api/apiDataHooks/transaction/useFetchFeeWorth'
